Brainnetome Atlas

Brain atlas for the Brainnetome cortical parcellation with 105 regions per hemisphere. Contains 2D polygon geometry for ggseg::geom_brain().

Usage

brainnetome()

Value

A ggseg.formats::ggseg_atlas object (cortical).

References

Fan L, Li H, Zhuo J, Zhang Y, Wang J, Chen L, Yang Z, Chu C, Xie S, Laird AR, Fox PT, Eickhoff SB, Yu C, Jiang T (2016). The Human Brainnetome Atlas: A New Brain Atlas Based on Connectional Architecture. Cerebral Cortex, 26(8):3508-3526. doi:10.1093/cercor/bhw157

See also

Examples

brainnetome()
#> 
#> ── brainnetome ggseg atlas ─────────────────────────────────────────────────────
#> Type: cortical
#> Regions: 210
#> Hemispheres: left, right
#> Views: inferior, lateral, medial, superior
#> Palette: ✔
#> Rendering: ✔ ggseg
#> ✔ ggseg3d (vertices)
#> ────────────────────────────────────────────────────────────────────────────────
#> # A tibble: 210 × 3
#>     hemi  region        label           
#>     <chr> <chr>         <chr>           
#>   1 left  A8m_L         lh_A8m_L        
#>   2 left  A8dl_L        lh_A8dl_L       
#>   3 left  A9l_L         lh_A9l_L        
#>   4 left  A6dl_L        lh_A6dl_L       
#>   5 left  A6m_L         lh_A6m_L        
#>   6 left  A9m_L         lh_A9m_L        
#>   7 left  A10m_L        lh_A10m_L       
#>   8 left  A9/46d_L      lh_A9-46d_L     
#>   9 left  IFJ_L         lh_IFJ_L        
#>  10 left  A46_L         lh_A46_L        
#>  11 left  A9/46v_L      lh_A9-46v_L     
#>  12 left  A8vl_L        lh_A8vl_L       
#>  13 left  A6vl_L        lh_A6vl_L       
#>  14 left  A10l_L        lh_A10l_L       
#>  15 left  A44d_L        lh_A44d_L       
#>  16 left  IFS_L         lh_IFS_L        
#>  17 left  A45c_L        lh_A45c_L       
#>  18 left  A45r_L        lh_A45r_L       
#>  19 left  A44op_L       lh_A44op_L      
#>  20 left  A44v_L        lh_A44v_L       
#>  21 left  A14m_L        lh_A14m_L       
#>  22 left  A12/47o_L     lh_A12-47o_L    
#>  23 left  A11l_L        lh_A11l_L       
#>  24 left  A11m_L        lh_A11m_L       
#>  25 left  A13_L         lh_A13_L        
#>  26 left  A12/47l_L     lh_A12-47l_L    
#>  27 left  A4hf_L        lh_A4hf_L       
#>  28 left  A6cdl_L       lh_A6cdl_L      
#>  29 left  A4ul_L        lh_A4ul_L       
#>  30 left  A4t_L         lh_A4t_L        
#>  31 left  A4tl_L        lh_A4tl_L       
#>  32 left  A6cvl_L       lh_A6cvl_L      
#>  33 left  A1/2/3ll_L    lh_A1-2-3ll_L   
#>  34 left  A4ll_L        lh_A4ll_L       
#>  35 left  A38m_L        lh_A38m_L       
#>  36 left  A41/42_L      lh_A41-42_L     
#>  37 left  TE1.0/TE1.2_L lh_TE1.0-TE1.2_L
#>  38 left  A22c_L        lh_A22c_L       
#>  39 left  A38l_L        lh_A38l_L       
#>  40 left  A22r_L        lh_A22r_L       
#>  41 left  A21c_L        lh_A21c_L       
#>  42 left  A21r_L        lh_A21r_L       
#>  43 left  A37dl_L       lh_A37dl_L      
#>  44 left  aSTS_L        lh_aSTS_L       
#>  45 left  A20iv_L       lh_A20iv_L      
#>  46 left  A37elv_L      lh_A37elv_L     
#>  47 left  A20r_L        lh_A20r_L       
#>  48 left  A20il_L       lh_A20il_L      
#>  49 left  A37vl_L       lh_A37vl_L      
#>  50 left  A20cl_L       lh_A20cl_L      
#>  51 left  A20cv_L       lh_A20cv_L      
#>  52 left  A20rv_L       lh_A20rv_L      
#>  53 left  A37mv_L       lh_A37mv_L      
#>  54 left  A37lv_L       lh_A37lv_L      
#>  55 left  A35/36r_L     lh_A35-36r_L    
#>  56 left  A35/36c_L     lh_A35-36c_L    
#>  57 left  TL_L          lh_TL_L         
#>  58 left  A28/34_L      lh_A28-34_L     
#>  59 left  TI_L          lh_TI_L         
#>  60 left  TH_L          lh_TH_L         
#>  61 left  rpSTS_L       lh_rpSTS_L      
#>  62 left  cpSTS_L       lh_cpSTS_L      
#>  63 left  A7r_L         lh_A7r_L        
#>  64 left  A7c_L         lh_A7c_L        
#>  65 left  A5l_L         lh_A5l_L        
#>  66 left  A7pc_L        lh_A7pc_L       
#>  67 left  A7ip_L        lh_A7ip_L       
#>  68 left  A39c_L        lh_A39c_L       
#>  69 left  A39rd_L       lh_A39rd_L      
#>  70 left  A40rd_L       lh_A40rd_L      
#>  71 left  A40c_L        lh_A40c_L       
#>  72 left  A39rv_L       lh_A39rv_L      
#>  73 left  A40rv_L       lh_A40rv_L      
#>  74 left  A7m_L         lh_A7m_L        
#>  75 left  A5m_L         lh_A5m_L        
#>  76 left  dmPOS_L       lh_dmPOS_L      
#>  77 left  A31_L         lh_A31_L        
#>  78 left  A1/2/3ulhf_L  lh_A1-2-3ulhf_L 
#>  79 left  A1/2/3tonIa_L lh_A1-2-3tonIa_L
#>  80 left  A2_L          lh_A2_L         
#>  81 left  A1/2/3tru_L   lh_A1-2-3tru_L  
#>  82 left  G_L           lh_G_L          
#>  83 left  vIa_L         lh_vIa_L        
#>  84 left  dIa_L         lh_dIa_L        
#>  85 left  vId/vIg_L     lh_vId-vIg_L    
#>  86 left  dIg_L         lh_dIg_L        
#>  87 left  dId_L         lh_dId_L        
#>  88 left  A23d_L        lh_A23d_L       
#>  89 left  A24rv_L       lh_A24rv_L      
#>  90 left  A32p_L        lh_A32p_L       
#>  91 left  A23v_L        lh_A23v_L       
#>  92 left  A24cd_L       lh_A24cd_L      
#>  93 left  A23c_L        lh_A23c_L       
#>  94 left  A32sg_L       lh_A32sg_L      
#>  95 left  cLinG_L       lh_cLinG_L      
#>  96 left  rCunG_L       lh_rCunG_L      
#>  97 left  cCunG_L       lh_cCunG_L      
#>  98 left  rLinG_L       lh_rLinG_L      
#>  99 left  vmPOS_L       lh_vmPOS_L      
#> 100 left  mOccG_L       lh_mOccG_L      
#> 101 left  V5/MT+_L      lh_V5-MT+_L     
#> 102 left  OPC_L         lh_OPC_L        
#> 103 left  iOccG_L       lh_iOccG_L      
#> 104 left  msOccG_L      lh_msOccG_L     
#> 105 left  lsOccG_L      lh_lsOccG_L     
#> 106 right A8m_R         rh_A8m_R        
#> 107 right A8dl_R        rh_A8dl_R       
#> 108 right A9l_R         rh_A9l_R        
#> 109 right A6dl_R        rh_A6dl_R       
#> 110 right A6m_R         rh_A6m_R        
#> 111 right A9m_R         rh_A9m_R        
#> 112 right A10m_R        rh_A10m_R       
#> 113 right A9/46d_R      rh_A9-46d_R     
#> 114 right IFJ_R         rh_IFJ_R        
#> 115 right A46_R         rh_A46_R        
#> 116 right A9/46v_R      rh_A9-46v_R     
#> 117 right A8vl_R        rh_A8vl_R       
#> 118 right A6vl_R        rh_A6vl_R       
#> 119 right A10l_R        rh_A10l_R       
#> 120 right A44d_R        rh_A44d_R       
#> 121 right IFS_R         rh_IFS_R        
#> 122 right A45c_R        rh_A45c_R       
#> 123 right A45r_R        rh_A45r_R       
#> 124 right A44op_R       rh_A44op_R      
#> 125 right A44v_R        rh_A44v_R       
#> 126 right A14m_R        rh_A14m_R       
#> 127 right A12/47o_R     rh_A12-47o_R    
#> 128 right A11l_R        rh_A11l_R       
#> 129 right A11m_R        rh_A11m_R       
#> 130 right A13_R         rh_A13_R        
#> 131 right A12/47l_R     rh_A12-47l_R    
#> 132 right A4hf_R        rh_A4hf_R       
#> 133 right A6cdl_R       rh_A6cdl_R      
#> 134 right A4ul_R        rh_A4ul_R       
#> 135 right A4t_R         rh_A4t_R        
#> 136 right A4tl_R        rh_A4tl_R       
#> 137 right A6cvl_R       rh_A6cvl_R      
#> 138 right A1/2/3ll_R    rh_A1-2-3ll_R   
#> 139 right A4ll_R        rh_A4ll_R       
#> 140 right A38m_R        rh_A38m_R       
#> 141 right A41/42_R      rh_A41-42_R     
#> 142 right TE1.0/TE1.2_R rh_TE1.0-TE1.2_R
#> 143 right A22c_R        rh_A22c_R       
#> 144 right A38l_R        rh_A38l_R       
#> 145 right A22r_R        rh_A22r_R       
#> 146 right A21c_R        rh_A21c_R       
#> 147 right A21r_R        rh_A21r_R       
#> 148 right A37dl_R       rh_A37dl_R      
#> 149 right aSTS_R        rh_aSTS_R       
#> 150 right A20iv_R       rh_A20iv_R      
#> 151 right A37elv_R      rh_A37elv_R     
#> 152 right A20r_R        rh_A20r_R       
#> 153 right A20il_R       rh_A20il_R      
#> 154 right A37vl_R       rh_A37vl_R      
#> 155 right A20cl_R       rh_A20cl_R      
#> 156 right A20cv_R       rh_A20cv_R      
#> 157 right A20rv_R       rh_A20rv_R      
#> 158 right A37mv_R       rh_A37mv_R      
#> 159 right A37lv_R       rh_A37lv_R      
#> 160 right A35/36r_R     rh_A35-36r_R    
#> 161 right A35/36c_R     rh_A35-36c_R    
#> 162 right TL_R          rh_TL_R         
#> 163 right A28/34_R      rh_A28-34_R     
#> 164 right TI_R          rh_TI_R         
#> 165 right TH_R          rh_TH_R         
#> 166 right rpSTS_R       rh_rpSTS_R      
#> 167 right cpSTS_R       rh_cpSTS_R      
#> 168 right A7r_R         rh_A7r_R        
#> 169 right A7c_R         rh_A7c_R        
#> 170 right A5l_R         rh_A5l_R        
#> 171 right A7pc_R        rh_A7pc_R       
#> 172 right A7ip_R        rh_A7ip_R       
#> 173 right A39c_R        rh_A39c_R       
#> 174 right A39rd_R       rh_A39rd_R      
#> 175 right A40rd_R       rh_A40rd_R      
#> 176 right A40c_R        rh_A40c_R       
#> 177 right A39rv_R       rh_A39rv_R      
#> 178 right A40rv_R       rh_A40rv_R      
#> 179 right A7m_R         rh_A7m_R        
#> 180 right A5m_R         rh_A5m_R        
#> 181 right dmPOS_R       rh_dmPOS_R      
#> 182 right A31_R         rh_A31_R        
#> 183 right A1/2/3ulhf_R  rh_A1-2-3ulhf_R 
#> 184 right A1/2/3tonIa_R rh_A1-2-3tonIa_R
#> 185 right A2_R          rh_A2_R         
#> 186 right A1/2/3tru_R   rh_A1-2-3tru_R  
#> 187 right G_R           rh_G_R          
#> 188 right vIa_R         rh_vIa_R        
#> 189 right dIa_R         rh_dIa_R        
#> 190 right vId/vIg_R     rh_vId-vIg_R    
#> 191 right dIg_R         rh_dIg_R        
#> 192 right dId_R         rh_dId_R        
#> 193 right A23d_R        rh_A23d_R       
#> 194 right A24rv_R       rh_A24rv_R      
#> 195 right A32p_R        rh_A32p_R       
#> 196 right A23v_R        rh_A23v_R       
#> 197 right A24cd_R       rh_A24cd_R      
#> 198 right A23c_R        rh_A23c_R       
#> 199 right A32sg_R       rh_A32sg_R      
#> 200 right cLinG_R       rh_cLinG_R      
#> 201 right rCunG_R       rh_rCunG_R      
#> 202 right cCunG_R       rh_cCunG_R      
#> 203 right rLinG_R       rh_rLinG_R      
#> 204 right vmPOS_R       rh_vmPOS_R      
#> 205 right mOccG_R       rh_mOccG_R      
#> 206 right V5/MT+_R      rh_V5-MT+_R     
#> 207 right OPC_R         rh_OPC_R        
#> 208 right iOccG_R       rh_iOccG_R      
#> 209 right msOccG_R      rh_msOccG_R     
#> 210 right lsOccG_R      rh_lsOccG_R     
if (FALSE) plot(brainnetome()) # \dontrun{}